requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==25.9.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
